Moved some functions around, cleaned up pooled normal handling a bit
